Rename the ui project back to web
Making the application run on the web seem to be a little trikier
than expected. It does not yet have proper support for `nannou_egui`
and wgpu support on web browser is still bery apporximate. Meaning the
ported application won't run on mobile devices as it stands.

I suggest we keep a running web version of the application that will be
updated as the nannou web support improves. Later a `ui` folder will be
created and will contain the full running version of the application.